+++ ext/monitor/src/main/java/app/monitor/MonitorConfig.java
@@ -83,7 +83,7 @@ public class MonitorConfig {
         @NotNull
         @Min(0)
         @Property(name = "highDiskSizeThreshold")
-        public Long highDiskSizeThreshold = 50_000_000_000L;    // use 50G as default
+        public Long highDiskSizeThreshold = 50_000_000_000L;    // use 50G as default, kafka jmx doesn't provide disk limit, usually in cloud/kube env, pod pvc disk usage can be monitored by cloud monitoring as well
     }
 
     public static class KubeConfig {